# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MultiBook API
  version: 1.0.0
  description: Private API consumed by the MultiBook mobile applications.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/users/me:
    get:
      summary: Read and provision the authenticated user profile
      security:
        - firebaseBearer: []
      responses:
        '200':
          description: Verified Firebase user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid Firebase ID token
    patch:
      summary: Update the authenticated user profile
      security:
        - firebaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMyProfileRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { description: Missing or invalid Firebase ID token }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/users/me/role:
    put:
      summary: Select the authenticated user's marketplace role
      security:
        - firebaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [customer, provider]
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { description: Missing or invalid Firebase ID token }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/users/me/selected-business:
    put:
      summary: Select an owned business as the authenticated provider's active business
      security:
        - firebaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [business_id]
              properties:
                business_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Active business selected
          content:
            application/json:
              schema:
                type: object
                required: [selected_business_id]
                properties:
                  selected_business_id: { type: string, format: uuid }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/businesses:
    post:
      summary: Create a business owned by the authenticated provider
      security:
        - firebaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessRequest'
      responses:
        '201':
          description: Created complete business aggregate, directly decodable by Flutter BusinessModel
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessModelResponse' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      summary: List lightweight summaries of businesses owned by the authenticated provider
      security:
        - firebaseBearer: []
      responses:
        '200':
          description: Lightweight owned-business summaries for selectors, tabs and lists. Use GET /v1/businesses/{businessID} before opening the full editor.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/BusinessSummaryResponse' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /v1/discovery/businesses:
    get:
      summary: List public businesses for discovery
      description: Customer discovery feed used by Popular Near You and the default Other businesses lists. Supplying city filters the page to that city; omitting it returns the global ranking. Results are lightweight camelCase BusinessModel-compatible cards ordered by rating and review count.
      security:
        - firebaseBearer: []
      parameters:
        - name: type
          in: query
          required: true
          schema: { type: string, enum: [stays, services] }
        - name: city
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 30, default: 10 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Page of popular nearby businesses
          content:
            application/json:
              schema:
                type: object
                required: [items, nextOffset]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/BusinessSummaryResponse' }
                  nextOffset: { type: integer, minimum: 0 }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/discovery/search:
    get:
      summary: Search public businesses by name or city
      description: Text search for customer stay and service listings. Matching uses normalized name and city prefixes.
      security:
        - firebaseBearer: []
      parameters:
        - name: type
          in: query
          required: true
          schema: { type: string, enum: [stays, services] }
        - name: query
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 30, default: 20 }
      responses:
        '200':
          description: Matching discovery cards
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/BusinessSummaryResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/development/seed/stays:
    post:
      summary: Seed demo stay businesses in development only
      description: Provider-only and available only when APP_ENV is development. Demo cover media is a direct HTTPS URL, not a Firebase Storage path.
      security: [{ firebaseBearer: [] }]
      responses:
        '201':
          description: Demo stays created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DevelopmentSeedResult' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /v1/development/seed/services:
    post:
      summary: Seed demo service businesses in development only
      description: Provider-only and available only when APP_ENV is development. Demo cover media is a direct HTTPS URL, not a Firebase Storage path.
      security: [{ firebaseBearer: [] }]
      responses:
        '201':
          description: Demo service businesses created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DevelopmentSeedResult' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /v1/businesses/{businessID}:
    parameters:
      - name: businessID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get the complete aggregate of an owned business for the provider editor
      security:
        - firebaseBearer: []
      responses:
        '200':
          description: Complete BusinessModel aggregate. The authenticated provider must own the business.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessModelResponse' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    put:
      summary: Replace an owned business aggregate
      description: Uses the existing Flutter BusinessModel.toMap() camelCase payload. This is a full replacement; fetch the detail endpoint before editing so nested stay or service fields are preserved.
      security:
        - firebaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBusinessRequest' }
      responses:
        '200':
          description: Replaced complete business aggregate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessModelResponse' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    patch:
      summary: Update editable fields of an owned business
      security:
        - firebaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateBusinessRequest' }
      responses:
        '200':
          description: Updated business
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Business' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      summary: Archive an owned business
      security:
        - firebaseBearer: []
      responses:
        '204': { description: Business archived }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /v1/businesses/{businessID}/location:
    parameters:
      - $ref: '#/components/parameters/BusinessID'
    put:
      summary: Create or replace the primary location of an owned business
      security: [{ firebaseBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertBusinessLocationRequest' }
      responses:
        '200': { description: Saved location, content: { application/json: { schema: { $ref: '#/components/schemas/BusinessLocation' } } } }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /v1/businesses/{businessID}/earnings/stream:
    parameters:
      - $ref: '#/components/parameters/BusinessID'
      - { name: startDate, in: query, required: true, schema: { type: string, format: date } }
      - { name: endDate, in: query, required: true, schema: { type: string, format: date } }
      - { name: utcOffsetMinutes, in: query, schema: { type: integer, minimum: -840, maximum: 840, default: 0 }, description: Client UTC offset used for calendar boundaries and daily grouping }
      - { name: staffId, in: query, schema: { type: string, format: uuid } }
    get:
      summary: Stream live earnings metrics for an owned business and date range
      security: [{ firebaseBearer: [] }]
      responses:
        '200':
          description: Server-Sent Events containing EarningsMetrics JSON in each earnings_metrics event
          content:
            text/event-stream:
              schema: { type: string }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /v1/conversations/unread-count:
    get:
      summary: Get the authenticated participant's aggregate unread message count
      security: [{ firebaseBearer: [] }]
      responses:
        '200':
          description: Unread message count
          content: { application/json: { schema: { type: object, required: [count], properties: { count: { type: integer, minimum: 0 } } } } }
...
components:
...
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  schemas:
...
    User:
      type: object
      required: [id, business_currency, created_at, updated_at]
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        role:
          type: string
          enum: [customer, provider, admin]
          nullable: true
        phone_e164:
          type: string
          nullable: true
          description: Phone number in E.164 format when supplied.
        avatar_storage_path:
          type: string
          nullable: true
          description: Firebase Storage path, not a download URL.
        country_code:
          type: string
          nullable: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        business_currency:
          type: string
          enum: [BAM, USD, EUR, CHF, GBP]
        selected_business_id: { type: string, nullable: true }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...